+++ QSymbolicsBase/src/QSymbolicsBase.jl
@@ -10,7 +10,10 @@ import TermInterface: istree, exprhead, operation, arguments, similarterm, metad
 using LinearAlgebra
 import LinearAlgebra: eigvecs
 
-import QuantumInterface: tensor, ⊗, basis, Basis, SpinBasis, projector,
+import QuantumInterface:
+    tensor, ⊗,
+    basis, Basis, SpinBasis, FockBasis,
+    projector,
     AbstractKet, AbstractOperator, AbstractSuperOperator
 
 #=
@@ -30,10 +33,13 @@ export SymQObj,QObj,
        X,Y,Z,σˣ,σʸ,σᶻ,Pm,Pp,σ₋,σ₊,
        H,CNOT,CPHASE,
        X1,X2,Y1,Y2,Z1,Z2,X₁,X₂,Y₁,Y₂,Z₁,Z₂,L0,L1,Lp,Lm,Lpi,Lmi,L₀,L₁,L₊,L₋,L₊ᵢ,L₋ᵢ,
+       vac,F₀,F0,F₁,F1,
+       N,n̂,Create,âꜛ,Destroy,â,
        SProjector,MixedState,IdentityOp,
        STensorKet,STensorOperator,SScaledKet,SScaledOperator,SAddKet,SAddOperator,
        HGate, XGate, YGate, ZGate, CPHASEGate, CNOTGate,
-       XBasisState, YBasisState, ZBasisState
+       XBasisState, YBasisState, ZBasisState,
+       NumberOp, CreateOp, DestroyOp
 
 function countmap(samples) # A simpler version of StatsBase.countmap, because StatsBase is slow to import
     counts = Dict{Any,Any}()