+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.1 - dev
 
+- Bosonic states basis and common states and operators.
 - Bumping `QSymbolicsBase` to 0.1.1.
 - Bumping `QSymbolicsOptics` to 0.1.1.
 - Added tests.