+++ QSymbolicsBase/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.1.1 - dev
 
+- Bosonic states basis and common states and operators.
 - Documentation and more symbol names available for the qubit base states.
 - `SApplyOp` implemented for superoperator acting on operator.
 - `STensorSuperOperator` implemented for tensor products of superoperators.