+++ QSymbolicsOptics/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## v0.1.1 - dev
 
+- Bosonic states basis and common states and operators.
 - Documentation and internal cleanup of the qubit bases.
 - `LazySuperTensor` implemented (should be moved to `QuantumOpticsBase`).